% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_outputs_by_journals}
\alias{cluster_outputs_by_journals}
\title{Group}
\usage{
cluster_outputs_by_journals(data)
}
\value{
The input data frame, with an added column \code{journal_id} assigning each
submission to a group representing a journal (or book). Any submissions that
could not be assigned to a group will have \code{journal_id = NA}.
}
\description{
Group
}
\note{
The current implementation is not 100% safe, as there is the possibility that
row IDs become large enough to be confused with ISSNs. We can fix this with
the addition of a non-numeric prefix to row ids.
}
\examples{
# Not done yet. Need to reconcile column names.

}
