% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_REF}
\alias{download_REF}
\title{Download data from the REF 2014 web site}
\usage{
download_REF(type = c("results", "outputs"), destination = tempdir())
}
\arguments{
\item{type}{Either 'outputs' or 'results'}

\item{destination}{Optional directory in which to store the downloaded
\code{.xlsx} file. If not supplied, it will be saved to a temporary directory}
}
\value{
Path to the downloaded/unzipped \code{.xlsx} file of REF outputs
}
\description{
This helper function will download and unzip an Excel (\code{.xlsx}) file
containing either the full results or the \emph{Research outputs (REF2)}
submission data for every institution and every unit of assessment in the
Research Excellence Framework 2014.
}
\details{
Alternatively you can download the data yourself manually from the
\href{https://results.ref.ac.uk/Results}{REF 2014 web site}.
Once downloaded, call \code{\link[readxl]{read_excel}} to import the data
from the spreadsheet into R.
}
\note{
Unfortunately there is no way to read Excel files from an \code{\link{unz}}
connection, so it is necessary to save the file to disk. However, if you
leave the \code{destination} to the default \code{tempdir()} then you should
not need to worry too much about cleaning up the \code{.xlsx} files
afterwards.
}
\seealso{
\code{\link[readxl]{read_excel}}, to import downloaded data into R
}
