% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_submissions}
\alias{download_submissions}
\title{Download submissions data from the REF 2014 web site}
\usage{
download_submissions(uoa = 1:36, destination = ".")
}
\arguments{
\item{uoa}{integer vector of unit(s) of assessment, from 1--36. Downloads all UoAs by default}

\item{destination}{directory in which to store the Excel spreadsheet(s).
Uses the current working directory by default}
}
\value{
A character vector of file paths to the downloaded \code{.xlsx} spreadsheets
}
\description{
Use this function to download an Excel spreadsheet from the REF 2014 web site
describing institutional submissions and results for a particular unit of
assessment (UoA). See Details. You probably don't want this function---it's
more efficient to use \code{\link{download_REF}('outputs')} instead.
}
\details{
A unit of assessment is an academic subject area, enumerated as follows.
See \href{https://results.ref.ac.uk/DownloadSubmissions/SelectUoa}{online}
for more information.
\enumerate{
\item Clinical Medicine
\item Public Health, Health Services and Primary Care
\item Allied Health Professions, Dentistry, Nursing and Pharmacy
\item Psychology, Psychiatry and Neuroscience
\item Biological Sciences
\item Agriculture, Veterinary and Food Science
\item Earth Systems and Environmental Sciences
\item Chemistry
\item Physics
\item Mathematical Sciences
\item Computer Science and Informatics
\item Aeronautical, Mechanical, Chemical and Manufacturing Engineering
\item Electrical and Electronic Engineering, Metallurgy and Materials
\item Civil and Construction Engineering
\item General Engineering
\item Architecture, Built Environment and Planning
\item Geography, Environmental Studies and Archaeology
\item Economics and Econometrics
\item Business and Management Studies
\item Law
\item Politics and International Studies
\item Social Work and Social Policy
\item Sociology
\item Anthropology and Development Studies
\item Education
\item Sport and Exercise Sciences, Leisure and Tourism
\item Area Studies
\item Modern Languages and Linguistics
\item English Language and Literature
\item History
\item Classics
\item Philosophy
\item Theology and Religious Studies
\item Art and Design: History, Practice and Theory
\item Music, Drama, Dance and Performing Arts
\item Communication, Cultural and Media Studies, Library and Information Management
}
}
