% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6EM.R
\docType{class}
\name{EM}
\alias{EM}
\title{Expectation-maximization algorithm for ecological REF data}
\format{An \code{\link{R6Class}} generator object}
\usage{
EM
}
\description{
Based on an R6 class.
}
\section{Methods}{

\describe{
  \item{\code{example_method(parameter_1 = 3)}}{This method uses \code{parameter_1} to...}
}
}

\examples{
# See strength.Rmd to generate the dataset.
\dontrun{
trials <- select(dataset, inst, journal, ntrials) \%>\% mutate_if(is.factor, as.character)
totals <- dataset \%>\% group_by(inst) \%>\% summarise(total = sum(success))

set.seed(1)
test <- EM$new(trials, totals, 2)
test$run(100)
test$iterations
}

}
\keyword{data}
