% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{get_top_journals}
\alias{get_top_journals}
\title{List the most popular journals}
\usage{
get_top_journals(outputs, min_articles = 20)
}
\arguments{
\item{outputs}{data from \code{\link{cluster_outputs_by_journals}}}

\item{min_articles}{'top' journals must have at least this many articles each}
}
\description{
Enumerate the journals, books and reports with the most REF submissions.
Any volumes with fewer than `min_articles` submissions will be aggregated
into a category called 'Other'.
}
\examples{
math_outputs <- subset(tidy_outputs(), uoa_name == 'Mathematical Sciences')
math_outputs <- cluster_outputs_by_journals(math_outputs)
get_top_journals(math_outputs)

}
