% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outputs}
\alias{outputs}
\title{Research Excellence Framework 2014 results}
\format{A tibble with 190,962 rows and 31 columns}
\source{
\url{https://results.ref.ac.uk/DownloadResults}
}
\usage{
outputs
}
\description{
This is a large data set and could possibly do with slimming down
}
\keyword{datasets}
