% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_outputs}
\alias{read_outputs}
\alias{read_profiles}
\title{Import REF data from an Excel spreadsheet}
\usage{
read_outputs(filename)

read_profiles(filename)
}
\arguments{
\item{filename}{Path to the Excel spreadsheet}
}
\description{
Extract data on institutional outputs and results from files downloaded from
the REF 2014 web site. Each file should correspond to a particular unit of
assessment (UoA). If you don't have the files already on your computer, you
can download them using \code{\link{download_uoa}}.
}
\details{
This is a light wrapper around \code{\link[readxl]{read_excel}}.
}
\seealso{
\code{\link{download_uoa}}, to download the spreadsheet files
}
