% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results}
\alias{results}
\title{Research output (REF2) submission data}
\format{A tibble with 7,644 rows and 16 columns}
\source{
\url{https://results.ref.ac.uk/DownloadSubmissions}
}
\usage{
results
}
\description{
Research output (REF2) submission data
}
\keyword{datasets}
