% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{standardise_volume_title}
\alias{standardise_volume_title}
\title{Convert a title to lower case, and remove spaces, punctuation and accents}
\usage{
standardise_volume_title(title)
}
\arguments{
\item{title}{A character vector of volume titles}
}
\value{
A character vector, the same length as the input
}
\description{
Convert a title to lower case, and remove spaces, punctuation and accents
}
