% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tidy_outputs}
\alias{tidy_outputs}
\title{Tidy the output data}
\usage{
tidy_outputs(outputs = ref2014::outputs)
}
\arguments{
\item{outputs}{A dataset of REF outputs. Defaults to whole REF2014 dataset}
}
\value{
The \code{outputs} data with column names and serial numbers sanitized
}
\description{
This is a utility function that performs three useful tasks on a REF outputs
dataset, making it more amenable to later analysis.
\itemize{
\item coerce column names to snake case (replacing spaces with underscores)
\item standardize ISBNs and ISSNs by removing hyphens and spaces
\item add a \code{volume_std} column representing a standardised journal/volume title
}
}
\examples{
# Extract Economics submissions data
econ <- grep('^Economics', UoA)
subset(tidy_outputs(outputs), uoa_number == econ)

}
\seealso{
\code{\link{standardise_volume_title}}
}
