% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UoA}
\alias{UoA}
\title{Units of Assessment}
\format{A character vector with 36 elements, giving the names of the UoAs}
\source{
\url{https://www.ref.ac.uk/2014/}
}
\usage{
UoA
}
\description{
The names of the subject-based units of assessment (UoAs) in the 2014 REF.
\enumerate{
\item Clinical Medicine
\item Public Health, Health Services and Primary Care
\item Allied Health Professions, Dentistry, Nursing and Pharmacy
\item Psychology, Psychiatry and Neuroscience
\item Biological Sciences
\item Agriculture, Veterinary and Food Science
\item Earth Systems and Environmental Sciences
\item Chemistry
\item Physics
\item Mathematical Sciences
\item Computer Science and Informatics
\item Aeronautical, Mechanical, Chemical and Manufacturing Engineering
\item Electrical and Electronic Engineering, Metallurgy and Materials
\item Civil and Construction Engineering
\item General Engineering
\item Architecture, Built Environment and Planning
\item Geography, Environmental Studies and Archaeology
\item Economics and Econometrics
\item Business and Management Studies
\item Law
\item Politics and International Studies
\item Social Work and Social Policy
\item Sociology
\item Anthropology and Development Studies
\item Education
\item Sport and Exercise Sciences, Leisure and Tourism
\item Area Studies
\item Modern Languages and Linguistics
\item English Language and Literature
\item History
\item Classics
\item Philosophy
\item Theology and Religious Studies
\item Art and Design: History, Practice and Theory
\item Music, Drama, Dance and Performing Arts
\item Communication, Cultural and Media Studies, Library and Information Management
}
}
\keyword{datasets}
