% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref2014.R
\docType{package}
\name{ref2014}
\alias{ref2014}
\alias{ref2014-package}
\title{REF 2014 package}
\description{
Access to data from the Research Excellence Framework 2014
}
\details{
See \url{https://github.com/Selbosh/ref2014} for more info.
}
